[bugfix] Set body upon NewRequestWithContext call
Due to the fact that `httpc` uses `bytes.Buffer` as the underlying `io.ReadCloser` for the request
body, the Content-Length will be set automatically if the body is already provided to `http.NewRequestWithContext`.

More importantly, the `GetBody` function will be set by `net/http`, thus making sure that the client
can follow redirects from 307 and 308 status codes.

Closes #32