[bugfix] Properly set request.GetBody for redirect handling (#34)
Even though the tests were successful for the changes leading to 1.0.17, the redirect would still not set the body correctly. I used an actual HTTP server as a test bed to verify this.

These changes work and are more or less copied from the net/http packages NewRequestWithContext function.

In the long run, we'll have to improve the test suite. It should have shown the error.

---------

Co-authored-by: Lennart Elsen <lelsen@open-systems.com>